requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
inflate64==0.3.1
install==1.3.5
multidict==6.0.4
multivolumefile==0.2.3
psutil==5.9.5
py7zr==0.20.5
pybcj==1.0.1
pycryptodomex==3.18.0
pyppmd==1.0.0
PyYAML==6.0
pyzstd==0.15.7
texttable==1.6.7
yarl==1.9.2
